refactor: migrate layer2 to layer1
