+++ contracts/package.json
@@ -53,12 +53,9 @@
     "test:scheduledBalanceAdjustments": "npx hardhat test test/unitTests/layer_2/scheduledTasks/scheduledBalanceAdjustments/scheduledBalanceAdjustments.test.ts",
     "test:scheduledTasks": "npx hardhat test test/unitTests/layer_2/scheduledTasks/scheduledTasks/scheduledTasks.test.ts",
     "test:cap": "npx hardhat test test/unitTests/layer_1/cap/cap.test.ts",
-    "test:Cap": "npx hardhat test test/unitTests/layer_2/cap/cap.test.ts",
     "test:securityUSA": "npx hardhat test test/unitTests/layer_3/securityUSA/securityUSA.test.ts",
     "test:lock": "npx hardhat test test/unitTests/layer_1/lock/lock.test.ts",
-    "test:Lock": "npx hardhat test test/unitTests/layer_2/lock/lock.test.ts",
     "test:hold": "npx hardhat test test/unitTests/layer_1/hold/hold.test.ts",
-    "test:Hold": "npx hardhat test test/unitTests/layer_2/hold/hold.test.ts",
     "test:diamondCutManager": "npx hardhat test test/unitTests/resolver/diamondCutManager.test.ts",
     "test:protectedPartitions": "NETWORK=hardhat npx hardhat test test/unitTests/layer_1/protectedPartitions/protectedPartitions.test.ts",
     "test:timeTravel": "NETWORK=hardhat npx hardhat test test/unitTests/timeTravel/timeTravel.test.ts",
