+++ contracts/test/unitTests/layer_1/hold/hold.test.ts
@@ -206,7 +206,11 @@
 import { expect } from 'chai'
 import { ethers } from 'hardhat'
 import { SignerWithAddress } from '@nomiclabs/hardhat-ethers/signers.js'
-import { takeSnapshot, time } from '@nomicfoundation/hardhat-network-helpers'
+import {
+    takeSnapshot,
+    loadFixture,
+    time,
+} from '@nomicfoundation/hardhat-network-helpers'
 import { SnapshotRestorer } from '@nomicfoundation/hardhat-network-helpers/src/helpers/takeSnapshot'
 import { isinGenerator } from '@thomaschaplin/isin-generator'
 import {
@@ -221,6 +225,10 @@ import {
     TimeTravel,
     Kyc,
     SsiManagement,
+    Equity,
+    AdjustBalances,
+    Cap,
+    AccessControl,
 } from '@typechain'
 import {
     PAUSER_ROLE,
@@ -229,6 +237,9 @@ import {
     CONTROL_LIST_ROLE,
     KYC_ROLE,
     SSI_MANAGER_ROLE,
+    CORPORATE_ACTION_ROLE,
+    ADJUSTMENT_BALANCE_ROLE,
+    CAP_ROLE,
     MAX_UINT256,
     deployEquityFromFactory,
     Rbac,
@@ -246,9 +257,22 @@ const _DEFAULT_PARTITION =
     '0x0000000000000000000000000000000000000000000000000000000000000001'
 const _WRONG_PARTITION =
     '0x0000000000000000000000000000000000000000000000000000000000000321'
+const _PARTITION_ID_1 =
+    '0x0000000000000000000000000000000000000000000000000000000000000001'
+const _PARTITION_ID_2 =
+    '0x0000000000000000000000000000000000000000000000000000000000000002'
 const _AMOUNT = 1000
 const _DATA = '0x1234'
+const maxSupply_Original = 1000000 * _AMOUNT
+const maxSupply_Partition_1_Original = 50000 * _AMOUNT
+const maxSupply_Partition_2_Original = 0
+const ONE_SECOND = 1
 const EMPTY_VC_ID = EMPTY_STRING
+const balanceOf_A_Original = [10 * _AMOUNT, 100 * _AMOUNT]
+const balanceOf_B_Original = [20 * _AMOUNT, 200 * _AMOUNT]
+const adjustFactor = 253
+const adjustDecimals = 2
+const decimals_Original = 6
 let holdIdentifier: any
 
 describe('Hold Tests', () => {
@@ -275,6 +299,10 @@ describe('Hold Tests', () => {
     let timeTravelFacet: TimeTravel
     let kycFacet: Kyc
     let ssiManagementFacet: SsiManagement
+    let equityFacet: Equity
+    let accessControlFacet: AccessControl
+    let capFacet: Cap
+    let adjustBalancesFacet: AdjustBalances
 
     const ONE_YEAR_IN_SECONDS = 365 * 24 * 60 * 60
     let currentTimestamp = 0
@@ -284,56 +312,14 @@ describe('Hold Tests', () => {
 
     let hold: any
 
-    before(async () => {
-        snapshot = await takeSnapshot()
-        // mute | mock console.log
-        console.log = () => {}
-        // eslint-disable-next-line @typescript-eslint/no-extra-semi
-        ;[signer_A, signer_B, signer_C, signer_D, signer_E] =
-            await ethers.getSigners()
-        account_A = signer_A.address
-        account_B = signer_B.address
-        account_C = signer_C.address
-        account_D = signer_D.address
-        account_E = signer_E.address
-
-        const { deployer, ...deployedContracts } =
-            await deployAtsFullInfrastructure(
-                await DeployAtsFullInfrastructureCommand.newInstance({
-                    signer: signer_A,
-                    useDeployed: false,
-                    timeTravelEnabled: true,
-                })
-            )
-
-        factory = deployedContracts.factory.contract
-        businessLogicResolver = deployedContracts.businessLogicResolver.contract
-    })
-
-    after(async () => {
-        await snapshot.restore()
-    })
-
-    afterEach(async () => {
-        await timeTravelFacet.resetSystemTimestamp()
-    })
-
-    async function deployAll(isMultiPartition: boolean) {
+    function set_initRbacs(): Rbac[] {
         const rbacIssuer: Rbac = {
             role: ISSUER_ROLE,
             members: [account_B],
         }
-        const rbacController: Rbac = {
-            role: CONTROLLER_ROLE,
-            members: [account_C],
-        }
         const rbacPausable: Rbac = {
             role: PAUSER_ROLE,
             members: [account_D],
-        }
-        const rbacControlList: Rbac = {
-            role: CONTROL_LIST_ROLE,
-            members: [account_E],
         }
         const rbacKYC: Rbac = {
             role: KYC_ROLE,
@@ -343,46 +329,30 @@ describe('Hold Tests', () => {
             role: SSI_MANAGER_ROLE,
             members: [account_A],
         }
-        const init_rbacs: Rbac[] = [
+        const rbacCorporateAction: Rbac = {
+            role: CORPORATE_ACTION_ROLE,
+            members: [account_B],
+        }
+        const rbacControlList: Rbac = {
+            role: CONTROL_LIST_ROLE,
+            members: [account_E],
+        }
+        const rbacController: Rbac = {
+            role: CONTROLLER_ROLE,
+            members: [account_C],
+        }
+        return [
             rbacIssuer,
-            rbacController,
             rbacPausable,
-            rbacControlList,
             rbacKYC,
             rbacSSI,
+            rbacCorporateAction,
+            rbacControlList,
+            rbacController,
         ]
+    }
 
-        diamond = await deployEquityFromFactory({
-            adminAccount: account_A,
-            isWhiteList: false,
-            isControllable: true,
-            arePartitionsProtected: false,
-            isMultiPartition: isMultiPartition,
-            name: 'TEST_Hold',
-            symbol: 'TAC',
-            decimals: 6,
-            isin: isinGenerator(),
-            votingRight: false,
-            informationRight: false,
-            liquidationRight: false,
-            subscriptionRight: true,
-            conversionRight: true,
-            redemptionRight: true,
-            putRight: false,
-            dividendRight: 1,
-            currency: '0x345678',
-            numberOfShares: MAX_UINT256,
-            nominalValue: 100,
-            regulationType: RegulationType.REG_D,
-            regulationSubType: RegulationSubType.REG_D_506_B,
-            countriesControlListType: true,
-            listOfCountries: 'ES,FR,CH',
-            info: 'nothing',
-            init_rbacs,
-            factory,
-            businessLogicResolver: businessLogicResolver.address,
-        })
-
+    async function setFacets({ diamond }: { diamond: ResolverProxy }) {
         holdFacet = await ethers.getContractAt(
             'Hold',
             diamond.address,
@@ -398,13 +368,14 @@ describe('Hold Tests', () => {
             diamond.address,
             signer_B
         )
-        controlListFacet = await ethers.getContractAt(
-            'ControlList',
+        kycFacet = await ethers.getContractAt('Kyc', diamond.address, signer_B)
+        ssiManagementFacet = await ethers.getContractAt(
+            'SsiManagement',
             diamond.address,
-            signer_E
+            signer_A
         )
-        erc20Facet = await ethers.getContractAt(
-            'ERC20',
+        equityFacet = await ethers.getContractAt(
+            'Equity',
             diamond.address,
             signer_A
         )
@@ -413,11 +384,27 @@ describe('Hold Tests', () => {
             diamond.address,
             signer_A
         )
-        kycFacet = await ethers.getContractAt('Kyc', diamond.address, signer_B)
-        ssiManagementFacet = await ethers.getContractAt(
-            'SsiManagement',
+        adjustBalancesFacet = await ethers.getContractAt(
+            'AdjustBalances',
             diamond.address,
             signer_A
+        )
+        capFacet = await ethers.getContractAt('Cap', diamond.address, signer_A)
+
+        accessControlFacet = await ethers.getContractAt(
+            'AccessControl',
+            diamond.address,
+            signer_A
+        )
+        erc20Facet = await ethers.getContractAt(
+            'ERC20',
+            diamond.address,
+            signer_A
+        )
+        controlListFacet = await ethers.getContractAt(
+            'ControlList',
+            diamond.address,
+            signer_E
         )
 
         await ssiManagementFacet.connect(signer_A).addIssuer(account_A)
@@ -450,6 +437,117 @@ describe('Hold Tests', () => {
             data: '0x',
         })
     }
+
+    async function deploySecurityFixtureMultiPartition() {
+        console.log('Multi partition deployed')
+        let init_rbacs: Rbac[] = set_initRbacs()
+
+        diamond = await deployEquityFromFactory({
+            adminAccount: account_A,
+            isWhiteList: false,
+            isControllable: true,
+            arePartitionsProtected: false,
+            isMultiPartition: true,
+            name: 'TEST_Lock',
+            symbol: 'TAC',
+            decimals: decimals_Original,
+            isin: isinGenerator(),
+            votingRight: false,
+            informationRight: false,
+            liquidationRight: false,
+            subscriptionRight: true,
+            conversionRight: true,
+            redemptionRight: true,
+            putRight: false,
+            dividendRight: 1,
+            currency: '0x345678',
+            numberOfShares: MAX_UINT256,
+            nominalValue: 100,
+            regulationType: RegulationType.REG_D,
+            regulationSubType: RegulationSubType.REG_D_506_B,
+            countriesControlListType: true,
+            listOfCountries: 'ES,FR,CH',
+            info: 'nothing',
+            init_rbacs,
+            factory,
+            businessLogicResolver: businessLogicResolver.address,
+        })
+
+        await setFacets({ diamond })
+    }
+
+    async function deploySecurityFixtureSinglePartition() {
+        console.log('Single partition deployed')
+
+        let init_rbacs: Rbac[] = set_initRbacs()
+
+        diamond = await deployEquityFromFactory({
+            adminAccount: account_A,
+            isWhiteList: false,
+            isControllable: true,
+            arePartitionsProtected: false,
+            isMultiPartition: false,
+            name: 'TEST_Lock',
+            symbol: 'TAC',
+            decimals: decimals_Original,
+            isin: isinGenerator(),
+            votingRight: false,
+            informationRight: false,
+            liquidationRight: false,
+            subscriptionRight: true,
+            conversionRight: true,
+            redemptionRight: true,
+            putRight: false,
+            dividendRight: 1,
+            currency: '0x345678',
+            numberOfShares: MAX_UINT256,
+            nominalValue: 100,
+            regulationType: RegulationType.REG_D,
+            regulationSubType: RegulationSubType.REG_D_506_B,
+            countriesControlListType: true,
+            listOfCountries: 'ES,FR,CH',
+            info: 'nothing',
+            init_rbacs,
+            factory,
+            businessLogicResolver: businessLogicResolver.address,
+        })
+
+        await setFacets({ diamond })
+    }
+
+    before(async () => {
+        snapshot = await takeSnapshot()
+        // mute | mock console.log
+        //console.log = () => {}
+        // eslint-disable-next-line @typescript-eslint/no-extra-semi
+        ;[signer_A, signer_B, signer_C, signer_D, signer_E] =
+            await ethers.getSigners()
+        account_A = signer_A.address
+        account_B = signer_B.address
+        account_C = signer_C.address
+        account_D = signer_D.address
+        account_E = signer_E.address
+
+        const { deployer, ...deployedContracts } =
+            await deployAtsFullInfrastructure(
+                await DeployAtsFullInfrastructureCommand.newInstance({
+                    signer: signer_A,
+                    useDeployed: false,
+                    timeTravelEnabled: true,
+                })
+            )
+
+        factory = deployedContracts.factory.contract
+        businessLogicResolver = deployedContracts.businessLogicResolver.contract
+    })
+
+    after(async () => {
+        await snapshot.restore()
+    })
+
+    afterEach(async () => {
+        await timeTravelFacet.resetSystemTimestamp()
+    })
 
     describe('Multi-partition disabled', () => {
         async function checkCreatedHold_expected(
@@ -525,7 +623,7 @@ describe('Hold Tests', () => {
                 holdId: 1,
             }
 
-            await deployAll(false)
+            await loadFixture(deploySecurityFixtureSinglePartition)
         })
 
         describe('Paused', () => {
@@ -782,59 +880,6 @@ describe('Hold Tests', () => {
                 )
             })
 
-            it('GIVEN a Token WHEN createHoldByPartition for wrong partition THEN transaction fails with InvalidPartition', async () => {
-                await deployAll(true)
-
-                console.log('createHoldByPartition')
-
-                await expect(
-                    holdFacet.createHoldByPartition(_WRONG_PARTITION, hold)
-                ).to.be.revertedWithCustomError(
-                    erc1410Facet,
-                    'InvalidPartition'
-                )
-
-                console.log('createHoldFromByPartition')
-
-                console.log('operatorCreateHoldByPartition')
-
-                await erc1410Facet
-                    .connect(signer_A)
-                    .authorizeOperator(account_B)
-
-                await expect(
-                    holdFacet
-                        .connect(signer_B)
-                        .operatorCreateHoldByPartition(
-                            _WRONG_PARTITION,
-                            account_A,
-                            hold,
-                            '0x'
-                        )
-                ).to.be.revertedWithCustomError(
-                    erc1410Facet,
-                    'InvalidPartition'
-                )
-
-                await erc1410Facet.connect(signer_A).revokeOperator(account_B)
-
-                console.log('controllerCreateHoldByPartition')
-
-                await expect(
-                    holdFacet
-                        .connect(signer_C)
-                        .controllerCreateHoldByPartition(
-                            _WRONG_PARTITION,
-                            account_A,
-                            hold,
-                            '0x'
-                        )
-                ).to.be.revertedWithCustomError(
-                    erc1410Facet,
-                    'InvalidPartition'
-                )
-            })
-
             it('GIVEN a Token WHEN createHoldByPartition passing empty escrow THEN transaction fails with ZeroAddressNotAllowed', async () => {
                 let hold_wrong = {
                     amount: _AMOUNT,
@@ -918,6 +963,7 @@ describe('Hold Tests', () => {
             })
 
             it('GIVEN a Token WHEN createHoldByPartition passing wrong expirationTimestamp THEN transaction fails with WrongExpirationTimestamp', async () => {
+                await timeTravelFacet.changeSystemTimestamp(currentTimestamp)
                 let wrongExpirationTimestamp = currentTimestamp - 1
 
                 let hold_wrong = {
@@ -927,8 +973,6 @@ describe('Hold Tests', () => {
                     to: ADDRESS_ZERO,
                     data: _DATA,
                 }
-
-                console.log('createHoldByPartition')
 
                 await expect(
                     holdFacet.createHoldByPartition(
@@ -1486,6 +1530,677 @@ describe('Hold Tests', () => {
                     0,
                     0
                 )
+            })
+        })
+    })
+
+    describe('Multi-partition enabled', () => {
+        beforeEach(async () => {
+            await loadFixture(deploySecurityFixtureMultiPartition)
+        })
+
+        it('GIVEN a Token WHEN createHoldByPartition for wrong partition THEN transaction fails with InvalidPartition', async () => {
+            console.log('createHoldByPartition')
+
+            await expect(
+                holdFacet.createHoldByPartition(_WRONG_PARTITION, hold)
+            ).to.be.revertedWithCustomError(erc1410Facet, 'InvalidPartition')
+
+            console.log('createHoldFromByPartition')
+
+            console.log('operatorCreateHoldByPartition')
+
+            await erc1410Facet.connect(signer_A).authorizeOperator(account_B)
+
+            await expect(
+                holdFacet
+                    .connect(signer_B)
+                    .operatorCreateHoldByPartition(
+                        _WRONG_PARTITION,
+                        account_A,
+                        hold,
+                        '0x'
+                    )
+            ).to.be.revertedWithCustomError(erc1410Facet, 'InvalidPartition')
+
+            await erc1410Facet.connect(signer_A).revokeOperator(account_B)
+
+            console.log('controllerCreateHoldByPartition')
+
+            await expect(
+                holdFacet
+                    .connect(signer_C)
+                    .controllerCreateHoldByPartition(
+                        _WRONG_PARTITION,
+                        account_A,
+                        hold,
+                        '0x'
+                    )
+            ).to.be.revertedWithCustomError(erc1410Facet, 'InvalidPartition')
+        })
+
+        describe('Adjust balances', () => {
+            async function setPreBalanceAdjustment() {
+                // Granting Role to account C
+                accessControlFacet = accessControlFacet.connect(signer_A)
+                await accessControlFacet.grantRole(
+                    ADJUSTMENT_BALANCE_ROLE,
+                    account_C
+                )
+                await accessControlFacet.grantRole(ISSUER_ROLE, account_A)
+                await accessControlFacet.grantRole(CAP_ROLE, account_A)
+                await accessControlFacet.grantRole(CONTROLLER_ROLE, account_A)
+
+                // Using account C (with role)
+                adjustBalancesFacet = adjustBalancesFacet.connect(signer_C)
+                erc1410Facet = erc1410Facet.connect(signer_A)
+                capFacet = capFacet.connect(signer_A)
+
+                await capFacet.setMaxSupply(maxSupply_Original)
+                await capFacet.setMaxSupplyByPartition(
+                    _PARTITION_ID_1,
+                    maxSupply_Partition_1_Original
+                )
+                await capFacet.setMaxSupplyByPartition(
+                    _PARTITION_ID_2,
+                    maxSupply_Partition_2_Original
+                )
+
+                await erc1410Facet.issueByPartition({
+                    partition: _PARTITION_ID_1,
+                    tokenHolder: account_A,
+                    value: balanceOf_A_Original[0],
+                    data: '0x',
+                })
+                await erc1410Facet.issueByPartition({
+                    partition: _PARTITION_ID_2,
+                    tokenHolder: account_A,
+                    value: balanceOf_A_Original[1],
+                    data: '0x',
+                })
+                await erc1410Facet.issueByPartition({
+                    partition: _PARTITION_ID_1,
+                    tokenHolder: account_B,
+                    value: balanceOf_B_Original[0],
+                    data: '0x',
+                })
+                await erc1410Facet.issueByPartition({
+                    partition: _PARTITION_ID_2,
+                    tokenHolder: account_B,
+                    value: balanceOf_B_Original[1],
+                    data: '0x',
+                })
+            }
+
+            it('GIVEN a hold WHEN adjustBalances THEN hold amount gets updated succeeds', async () => {
+                await setPreBalanceAdjustment()
+
+                const balance_Before = await erc1410Facet.balanceOf(account_A)
+                const balance_Before_Partition_1 =
+                    await erc1410Facet.balanceOfByPartition(
+                        _PARTITION_ID_1,
+                        account_A
+                    )
+
+                // HOLD
+                holdFacet = holdFacet.connect(signer_A)
+                let hold = {
+                    amount: _AMOUNT,
+                    expirationTimestamp: dateToUnixTimestamp(
+                        '2030-01-01T00:00:01Z'
+                    ),
+                    escrow: account_B,
+                    to: ADDRESS_ZERO,
+                    data: '0x',
+                }
+
+                await holdFacet.createHoldByPartition(_PARTITION_ID_1, hold)
+
+                const hold_TotalAmount_Before =
+                    await holdFacet.getHeldAmountFor(account_A)
+                const hold_TotalAmount_Before_Partition_1 =
+                    await holdFacet.getHeldAmountForByPartition(
+                        _PARTITION_ID_1,
+                        account_A
+                    )
+                const hold_Before = await holdFacet.getHoldForByPartition(
+                    holdIdentifier
+                )
+
+                // adjustBalances
+                await adjustBalancesFacet.adjustBalances(
+                    adjustFactor,
+                    adjustDecimals
+                )
+
+                // scheduled two balance updates
+                equityFacet = equityFacet.connect(signer_B)
+
+                const balanceAdjustmentData = {
+                    executionDate: dateToUnixTimestamp(
+                        '2030-01-01T00:00:02Z'
+                    ).toString(),
+                    factor: adjustFactor,
+                    decimals: adjustDecimals,
+                }
+
+                const balanceAdjustmentData_2 = {
+                    executionDate: dateToUnixTimestamp(
+                        '2030-01-01T00:16:40Z'
+                    ).toString(),
+                    factor: adjustFactor,
+                    decimals: adjustDecimals,
+                }
+                await equityFacet.setScheduledBalanceAdjustment(
+                    balanceAdjustmentData
+                )
+                await equityFacet.setScheduledBalanceAdjustment(
+                    balanceAdjustmentData_2
+                )
+
+                // wait for first scheduled balance adjustment only
+                await timeTravelFacet.changeSystemTimestamp(
+                    dateToUnixTimestamp('2030-01-01T00:00:03Z')
+                )
+
+                const hold_TotalAmount_After =
+                    await holdFacet.getHeldAmountForAdjusted(account_A)
+                const hold_TotalAmount_After_Partition_1 =
+                    await holdFacet.getHeldAmountForByPartitionAdjusted(
+                        _PARTITION_ID_1,
+                        account_A
+                    )
+                const hold_After =
+                    await holdFacet.getHoldForByPartitionAdjusted(
+                        holdIdentifier
+                    )
+                const balance_After = await erc1410Facet.balanceOf(account_A)
+                const balance_After_Partition_1 =
+                    await erc1410Facet.balanceOfByPartition(
+                        _PARTITION_ID_1,
+                        account_A
+                    )
+
+                expect(hold_TotalAmount_After).to.be.equal(
+                    hold_TotalAmount_Before.mul(adjustFactor * adjustFactor)
+                )
+                expect(hold_TotalAmount_After_Partition_1).to.be.equal(
+                    hold_TotalAmount_Before_Partition_1.mul(
+                        adjustFactor * adjustFactor
+                    )
+                )
+                expect(balance_After).to.be.equal(
+                    balance_Before.sub(_AMOUNT).mul(adjustFactor * adjustFactor)
+                )
+                expect(hold_TotalAmount_After).to.be.equal(
+                    hold_TotalAmount_Before.mul(adjustFactor * adjustFactor)
+                )
+                expect(balance_After_Partition_1).to.be.equal(
+                    balance_Before_Partition_1
+                        .sub(_AMOUNT)
+                        .mul(adjustFactor * adjustFactor)
+                )
+                expect(hold_After.amount_).to.be.equal(
+                    hold_Before.amount_.mul(adjustFactor * adjustFactor)
+                )
+            })
+
+            it('GIVEN a hold WHEN adjustBalances THEN execute succeed', async () => {
+                await setPreBalanceAdjustment()
+                const balance_Before_A = await erc1410Facet.balanceOf(account_A)
+                const balance_Before_Partition_1_A =
+                    await erc1410Facet.balanceOfByPartition(
+                        _PARTITION_ID_1,
+                        account_A
+                    )
+                const balance_Before_C = await erc1410Facet.balanceOf(account_C)
+                const balance_Before_Partition_1_C =
+                    await erc1410Facet.balanceOfByPartition(
+                        _PARTITION_ID_1,
+                        account_C
+                    )
+
+                // HOLD TWICE
+                const currentTimestamp = (
+                    await ethers.provider.getBlock('latest')
+                ).timestamp
+
+                holdFacet = holdFacet.connect(signer_A)
+                let hold = {
+                    amount: _AMOUNT,
+                    expirationTimestamp: currentTimestamp + 10 * ONE_SECOND,
+                    escrow: account_B,
+                    to: ADDRESS_ZERO,
+                    data: '0x',
+                }
+
+                await holdFacet.createHoldByPartition(_PARTITION_ID_1, hold)
+
+                hold.expirationTimestamp = currentTimestamp + 100 * ONE_SECOND
+                await holdFacet.createHoldByPartition(_PARTITION_ID_1, hold)
+
+                const held_Amount_Before = await holdFacet.getHeldAmountFor(
+                    account_A
+                )
+                const held_Amount_Before_Partition_1 =
+                    await holdFacet.getHeldAmountForByPartition(
+                        _PARTITION_ID_1,
+                        account_A
+                    )
+
+                // adjustBalances
+                await adjustBalancesFacet.adjustBalances(
+                    adjustFactor,
+                    adjustDecimals
+                )
+
+                // EXECUTE HOLD
+                await holdFacet
+                    .connect(signer_B)
+                    .executeHoldByPartition(
+                        holdIdentifier,
+                        account_C,
+                        hold.amount * adjustFactor
+                    )
+
+                const balance_After_Execute_A = await erc1410Facet.balanceOf(
+                    account_A
+                )
+                const balance_After_Execute_Partition_1_A =
+                    await erc1410Facet.balanceOfByPartition(
+                        _PARTITION_ID_1,
+                        account_A
+                    )
+                const balance_After_Execute_C = await erc1410Facet.balanceOf(
+                    account_C
+                )
+                const balance_After_Execute_Partition_1_C =
+                    await erc1410Facet.balanceOfByPartition(
+                        _PARTITION_ID_1,
+                        account_C
+                    )
+                const held_Amount_After = await holdFacet.getHeldAmountFor(
+                    account_A
+                )
+                const held_Amount_After_Partition_1 =
+                    await holdFacet.getHeldAmountForByPartition(
+                        _PARTITION_ID_1,
+                        account_A
+                    )
+
+                expect(balance_After_Execute_A).to.be.equal(
+                    balance_Before_A.sub(_AMOUNT).sub(_AMOUNT).mul(adjustFactor)
+                )
+                expect(balance_After_Execute_C).to.be.equal(
+                    balance_Before_C.add(_AMOUNT).mul(adjustFactor)
+                )
+                expect(balance_After_Execute_Partition_1_A).to.be.equal(
+                    balance_Before_Partition_1_A
+                        .sub(_AMOUNT)
+                        .sub(_AMOUNT)
+                        .mul(adjustFactor)
+                )
+                expect(balance_After_Execute_Partition_1_C).to.be.equal(
+                    balance_Before_Partition_1_C.add(_AMOUNT).mul(adjustFactor)
+                )
+                expect(held_Amount_After).to.be.equal(
+                    held_Amount_Before.sub(_AMOUNT).mul(adjustFactor)
+                )
+                expect(held_Amount_After_Partition_1).to.be.equal(
+                    held_Amount_Before_Partition_1
+                        .sub(_AMOUNT)
+                        .mul(adjustFactor)
+                )
+                expect(
+                    balance_After_Execute_A.add(held_Amount_After)
+                ).to.be.equal(balance_Before_A.sub(_AMOUNT).mul(adjustFactor))
+                expect(
+                    balance_After_Execute_Partition_1_A.add(
+                        held_Amount_After_Partition_1
+                    )
+                ).to.be.equal(
+                    balance_Before_Partition_1_A.sub(_AMOUNT).mul(adjustFactor)
+                )
+            })
+
+            it('GIVEN a hold WHEN adjustBalances THEN release succeed', async () => {
+                await setPreBalanceAdjustment()
+                const balance_Before = await erc1410Facet.balanceOf(account_A)
+                const balance_Before_Partition_1 =
+                    await erc1410Facet.balanceOfByPartition(
+                        _PARTITION_ID_1,
+                        account_A
+                    )
+
+                // HOLD TWICE
+                const currentTimestamp = (
+                    await ethers.provider.getBlock('latest')
+                ).timestamp
+
+                holdFacet = holdFacet.connect(signer_A)
+                let hold = {
+                    amount: _AMOUNT,
+                    expirationTimestamp: currentTimestamp + 10 * ONE_SECOND,
+                    escrow: account_B,
+                    to: ADDRESS_ZERO,
+                    data: '0x',
+                }
+
+                await holdFacet.createHoldByPartition(_PARTITION_ID_1, hold)
+
+                hold.expirationTimestamp = currentTimestamp + 100 * ONE_SECOND
+                await holdFacet.createHoldByPartition(_PARTITION_ID_1, hold)
+
+                const held_Amount_Before = await holdFacet.getHeldAmountFor(
+                    account_A
+                )
+                const held_Amount_Before_Partition_1 =
+                    await holdFacet.getHeldAmountForByPartition(
+                        _PARTITION_ID_1,
+                        account_A
+                    )
+
+                // adjustBalances
+                await adjustBalancesFacet.adjustBalances(
+                    adjustFactor,
+                    adjustDecimals
+                )
+
+                // RELEASE HOLD
+                await holdFacet
+                    .connect(signer_B)
+                    .releaseHoldByPartition(
+                        holdIdentifier,
+                        hold.amount * adjustFactor
+                    )
+
+                const balance_After_Release = await erc1410Facet.balanceOf(
+                    account_A
+                )
+                const balance_After_Release_Partition_1 =
+                    await erc1410Facet.balanceOfByPartition(
+                        _PARTITION_ID_1,
+                        account_A
+                    )
+                const held_Amount_After = await holdFacet.getHeldAmountFor(
+                    account_A
+                )
+                const held_Amount_After_Partition_1 =
+                    await holdFacet.getHeldAmountForByPartition(
+                        _PARTITION_ID_1,
+                        account_A
+                    )
+
+                expect(balance_After_Release).to.be.equal(
+                    balance_Before.sub(_AMOUNT).mul(adjustFactor)
+                )
+                expect(balance_After_Release_Partition_1).to.be.equal(
+                    balance_Before_Partition_1.sub(_AMOUNT).mul(adjustFactor)
+                )
+                expect(held_Amount_After).to.be.equal(
+                    held_Amount_Before.sub(_AMOUNT).mul(adjustFactor)
+                )
+                expect(held_Amount_After_Partition_1).to.be.equal(
+                    held_Amount_Before_Partition_1
+                        .sub(_AMOUNT)
+                        .mul(adjustFactor)
+                )
+                expect(
+                    balance_After_Release.add(held_Amount_After)
+                ).to.be.equal(balance_Before.mul(adjustFactor))
+                expect(
+                    balance_After_Release_Partition_1.add(
+                        held_Amount_After_Partition_1
+                    )
+                ).to.be.equal(balance_Before_Partition_1.mul(adjustFactor))
+            })
+
+            it('GIVEN a hold WHEN adjustBalances THEN reclaim succeed', async () => {
+                await setPreBalanceAdjustment()
+                const balance_Before = await erc1410Facet.balanceOf(account_A)
+                const balance_Before_Partition_1 =
+                    await erc1410Facet.balanceOfByPartition(
+                        _PARTITION_ID_1,
+                        account_A
+                    )
+
+                // HOLD TWICE
+                const currentTimestamp = (
+                    await ethers.provider.getBlock('latest')
+                ).timestamp
+
+                holdFacet = holdFacet.connect(signer_A)
+                let hold = {
+                    amount: _AMOUNT,
+                    expirationTimestamp: currentTimestamp + ONE_SECOND,
+                    escrow: account_B,
+                    to: ADDRESS_ZERO,
+                    data: '0x',
+                }
+
+                await holdFacet.createHoldByPartition(_PARTITION_ID_1, hold)
+
+                hold.expirationTimestamp = currentTimestamp + 100 * ONE_SECOND
+                await holdFacet.createHoldByPartition(_PARTITION_ID_1, hold)
+
+                const held_Amount_Before = await holdFacet.getHeldAmountFor(
+                    account_A
+                )
+                const held_Amount_Before_Partition_1 =
+                    await holdFacet.getHeldAmountForByPartition(
+                        _PARTITION_ID_1,
+                        account_A
+                    )
+
+                // adjustBalances
+                await adjustBalancesFacet.adjustBalances(
+                    adjustFactor,
+                    adjustDecimals
+                )
+
+                // RECLAIM HOLD
+                await time.setNextBlockTimestamp(
+                    (await ethers.provider.getBlock('latest')).timestamp +
+                        2 * ONE_SECOND
+                )
+                await holdFacet.reclaimHoldByPartition(holdIdentifier)
+
+                const balance_After_Release = await erc1410Facet.balanceOf(
+                    account_A
+                )
+                const balance_After_Release_Partition_1 =
+                    await erc1410Facet.balanceOfByPartition(
+                        _PARTITION_ID_1,
+                        account_A
+                    )
+                const held_Amount_After = await holdFacet.getHeldAmountFor(
+                    account_A
+                )
+                const held_Amount_After_Partition_1 =
+                    await holdFacet.getHeldAmountForByPartition(
+                        _PARTITION_ID_1,
+                        account_A
+                    )
+
+                expect(balance_After_Release).to.be.equal(
+                    balance_Before.sub(_AMOUNT).mul(adjustFactor)
+                )
+                expect(balance_After_Release_Partition_1).to.be.equal(
+                    balance_Before_Partition_1.sub(_AMOUNT).mul(adjustFactor)
+                )
+                expect(held_Amount_After).to.be.equal(
+                    held_Amount_Before.sub(_AMOUNT).mul(adjustFactor)
+                )
+                expect(held_Amount_After_Partition_1).to.be.equal(
+                    held_Amount_Before_Partition_1
+                        .sub(_AMOUNT)
+                        .mul(adjustFactor)
+                )
+                expect(
+                    balance_After_Release.add(held_Amount_After)
+                ).to.be.equal(balance_Before.mul(adjustFactor))
+                expect(
+                    balance_After_Release_Partition_1.add(
+                        held_Amount_After_Partition_1
+                    )
+                ).to.be.equal(balance_Before_Partition_1.mul(adjustFactor))
+            })
+
+            it('GIVEN a hold WHEN adjustBalances THEN hold succeeds', async () => {
+                await setPreBalanceAdjustment()
+                const balance_Before = await erc1410Facet.balanceOf(account_A)
+                const balance_Before_Partition_1 =
+                    await erc1410Facet.balanceOfByPartition(
+                        _PARTITION_ID_1,
+                        account_A
+                    )
+
+                // HOLD BEFORE BALANCE ADJUSTMENT
+                const currentTimestamp = (
+                    await ethers.provider.getBlock('latest')
+                ).timestamp
+
+                holdFacet = holdFacet.connect(signer_A)
+                let hold = {
+                    amount: _AMOUNT,
+                    expirationTimestamp: currentTimestamp + 100 * ONE_SECOND,
+                    escrow: account_B,
+                    to: ADDRESS_ZERO,
+                    data: '0x',
+                }
+
+                await holdFacet.createHoldByPartition(_PARTITION_ID_1, hold)
+
+                const held_Amount_Before = await holdFacet.getHeldAmountFor(
+                    account_A
+                )
+                const held_Amount_Before_Partition_1 =
+                    await holdFacet.getHeldAmountForByPartition(
+                        _PARTITION_ID_1,
+                        account_A
+                    )
+
+                // adjustBalances
+                await adjustBalancesFacet.adjustBalances(
+                    adjustFactor,
+                    adjustDecimals
+                )
+
+                // HOLD AFTER BALANCE ADJUSTMENT
+                holdFacet = holdFacet.connect(signer_A)
+                await holdFacet.createHoldByPartition(_PARTITION_ID_1, hold)
+
+                const balance_After_Hold = await erc1410Facet.balanceOf(
+                    account_A
+                )
+                const balance_After_Hold_Partition_1 =
+                    await erc1410Facet.balanceOfByPartition(
+                        _PARTITION_ID_1,
+                        account_A
+                    )
+                const held_Amount_After = await holdFacet.getHeldAmountFor(
+                    account_A
+                )
+                const held_Amount_After_Partition_1 =
+                    await holdFacet.getHeldAmountForByPartition(
+                        _PARTITION_ID_1,
+                        account_A
+                    )
+
+                expect(balance_After_Hold).to.be.equal(
+                    balance_Before.sub(_AMOUNT).mul(adjustFactor).sub(_AMOUNT)
+                )
+                expect(balance_After_Hold_Partition_1).to.be.equal(
+                    balance_Before_Partition_1
+                        .sub(_AMOUNT)
+                        .mul(adjustFactor)
+                        .sub(_AMOUNT)
+                )
+                expect(held_Amount_After).to.be.equal(
+                    held_Amount_Before.mul(adjustFactor).add(_AMOUNT)
+                )
+                expect(held_Amount_After_Partition_1).to.be.equal(
+                    held_Amount_Before_Partition_1
+                        .mul(adjustFactor)
+                        .add(_AMOUNT)
+                )
+                expect(balance_After_Hold.add(held_Amount_After)).to.be.equal(
+                    balance_Before.mul(adjustFactor)
+                )
+                expect(
+                    balance_After_Hold_Partition_1.add(
+                        held_Amount_After_Partition_1
+                    )
+                ).to.be.equal(balance_Before_Partition_1.mul(adjustFactor))
+            })
+
+            it('GIVEN a hold WHEN adjustBalances and release THEN reset hold LABAF succeeds', async () => {
+                await setPreBalanceAdjustment()
+
+                // HOLD
+                const currentTimestamp = (
+                    await ethers.provider.getBlock('latest')
+                ).timestamp
+
+                holdFacet = holdFacet.connect(signer_A)
+                let hold = {
+                    amount: _AMOUNT,
+                    expirationTimestamp: currentTimestamp + 10 * ONE_SECOND,
+                    escrow: account_B,
+                    to: ADDRESS_ZERO,
+                    data: '0x',
+                }
+
+                await holdFacet.createHoldByPartition(_PARTITION_ID_1, hold)
+
+                const TotalHoldLABAF_Before =
+                    await adjustBalancesFacet.getTotalHeldLabafByPartition(
+                        _PARTITION_ID_1,
+                        account_A
+                    )
+
+                // adjustBalances
+                await adjustBalancesFacet.adjustBalances(
+                    adjustFactor,
+                    adjustDecimals
+                )
+
+                // HOLD AGAIN
+                hold.expirationTimestamp = currentTimestamp + 100 * ONE_SECOND
+                await holdFacet.createHoldByPartition(_PARTITION_ID_1, hold)
+
+                const HoldLABAF_Before =
+                    await adjustBalancesFacet.getHoldLabafByPartition(
+                        _PARTITION_ID_1,
+                        2,
+                        account_A
+                    )
+
+                // EXECUTE HOLD
+                await holdFacet
+                    .connect(signer_B)
+                    .executeHoldByPartition(
+                        holdIdentifier,
+                        account_C,
+                        hold.amount
+                    )
+
+                const TotalHoldLABAF_After =
+                    await adjustBalancesFacet.getTotalHeldLabafByPartition(
+                        _PARTITION_ID_1,
+                        account_A
+                    )
+                const HoldLABAF_After =
+                    await adjustBalancesFacet.getHoldLabafByPartition(
+                        _PARTITION_ID_1,
+                        2,
+                        account_A
+                    )
+
+                expect(HoldLABAF_Before).to.be.equal(HoldLABAF_After)
+                expect(TotalHoldLABAF_Before).to.be.equal(0)
+                expect(TotalHoldLABAF_After).to.be.equal(adjustFactor)
+                expect(HoldLABAF_Before).to.be.equal(adjustFactor)
             })
         })
     })
