--- contracts/test/unitTests/layer_2/lock/lock.test.ts
@@ -1,788 +1,0 @@
-/*
-                                 Apache License
-                           Version 2.0, January 2004
-                        http://www.apache.org/licenses/
-
-   TERMS AND CONDITIONS FOR USE, REPRODUCTION, AND DISTRIBUTION
-
-   1. Definitions.
-
-      "License" shall mean the terms and conditions for use, reproduction,
-      and distribution as defined by Sections 1 through 9 of this document.
-
-      "Licensor" shall mean the copyright owner or entity authorized by
-      the copyright owner that is granting the License.
-
-      "Legal Entity" shall mean the union of the acting entity and all
-      other entities that control, are controlled by, or are under common
-      control with that entity. For the purposes of this definition,
-      "control" means (i) the power, direct or indirect, to cause the
-      direction or management of such entity, whether by contract or
-      otherwise, or (ii) ownership of fifty percent (50%) or more of the
-      outstanding shares, or (iii) beneficial ownership of such entity.
-
-      "You" (or "Your") shall mean an individual or Legal Entity
-      exercising permissions granted by this License.
-
-      "Source" form shall mean the preferred form for making modifications,
-      including but not limited to software source code, documentation
-      source, and configuration files.
-
-      "Object" form shall mean any form resulting from mechanical
-      transformation or translation of a Source form, including but
-      not limited to compiled object code, generated documentation,
-      and conversions to other media types.
-
-      "Work" shall mean the work of authorship, whether in Source or
-      Object form, made available under the License, as indicated by a
-      copyright notice that is included in or attached to the work
-      (an example is provided in the Appendix below).
-
-      "Derivative Works" shall mean any work, whether in Source or Object
-      form, that is based on (or derived from) the Work and for which the
-      editorial revisions, annotations, elaborations, or other modifications
-      represent, as a whole, an original work of authorship. For the purposes
-      of this License, Derivative Works shall not include works that remain
-      separable from, or merely link (or bind by name) to the interfaces of,
-      the Work and Derivative Works thereof.
-
-      "Contribution" shall mean any work of authorship, including
-      the original version of the Work and any modifications or additions
-      to that Work or Derivative Works thereof, that is intentionally
-      submitted to Licensor for inclusion in the Work by the copyright owner
-      or by an individual or Legal Entity authorized to submit on behalf of
-      the copyright owner. For the purposes of this definition, "submitted"
-      means any form of electronic, verbal, or written communication sent
-      to the Licensor or its representatives, including but not limited to
-      communication on electronic mailing lists, source code control systems,
-      and issue tracking systems that are managed by, or on behalf of, the
-      Licensor for the purpose of discussing and improving the Work, but
-      excluding communication that is conspicuously marked or otherwise
-      designated in writing by the copyright owner as "Not a Contribution."
-
-      "Contributor" shall mean Licensor and any individual or Legal Entity
-      on behalf of whom a Contribution has been received by Licensor and
-      subsequently incorporated within the Work.
-
-   2. Grant of Copyright License. Subject to the terms and conditions of
-      this License, each Contributor hereby grants to You a perpetual,
-      worldwide, non-exclusive, no-charge, royalty-free, irrevocable
-      copyright license to reproduce, prepare Derivative Works of,
-      publicly display, publicly perform, sublicense, and distribute the
-      Work and such Derivative Works in Source or Object form.
-
-   3. Grant of Patent License. Subject to the terms and conditions of
-      this License, each Contributor hereby grants to You a perpetual,
-      worldwide, non-exclusive, no-charge, royalty-free, irrevocable
-      (except as stated in this section) patent license to make, have made,
-      use, offer to sell, sell, import, and otherwise transfer the Work,
-      where such license applies only to those patent claims licensable
-      by such Contributor that are necessarily infringed by their
-      Contribution(s) alone or by combination of their Contribution(s)
-      with the Work to which such Contribution(s) was submitted. If You
-      institute patent litigation against any entity (including a
-      cross-claim or counterclaim in a lawsuit) alleging that the Work
-      or a Contribution incorporated within the Work constitutes direct
-      or contributory patent infringement, then any patent licenses
-      granted to You under this License for that Work shall terminate
-      as of the date such litigation is filed.
-
-   4. Redistribution. You may reproduce and distribute copies of the
-      Work or Derivative Works thereof in any medium, with or without
-      modifications, and in Source or Object form, provided that You
-      meet the following conditions:
-
-      (a) You must give any other recipients of the Work or
-          Derivative Works a copy of this License; and
-
-      (b) You must cause any modified files to carry prominent notices
-          stating that You changed the files; and
-
-      (c) You must retain, in the Source form of any Derivative Works
-          that You distribute, all copyright, patent, trademark, and
-          attribution notices from the Source form of the Work,
-          excluding those notices that do not pertain to any part of
-          the Derivative Works; and
-
-      (d) If the Work includes a "NOTICE" text file as part of its
-          distribution, then any Derivative Works that You distribute must
-          include a readable copy of the attribution notices contained
-          within such NOTICE file, excluding those notices that do not
-          pertain to any part of the Derivative Works, in at least one
-          of the following places: within a NOTICE text file distributed
-          as part of the Derivative Works; within the Source form or
-          documentation, if provided along with the Derivative Works; or,
-          within a display generated by the Derivative Works, if and
-          wherever such third-party notices normally appear. The contents
-          of the NOTICE file are for informational purposes only and
-          do not modify the License. You may add Your own attribution
-          notices within Derivative Works that You distribute, alongside
-          or as an addendum to the NOTICE text from the Work, provided
-          that such additional attribution notices cannot be construed
-          as modifying the License.
-
-      You may add Your own copyright statement to Your modifications and
-      may provide additional or different license terms and conditions
-      for use, reproduction, or distribution of Your modifications, or
-      for any such Derivative Works as a whole, provided Your use,
-      reproduction, and distribution of the Work otherwise complies with
-      the conditions stated in this License.
-
-   5. Submission of Contributions. Unless You explicitly state otherwise,
-      any Contribution intentionally submitted for inclusion in the Work
-      by You to the Licensor shall be under the terms and conditions of
-      this License, without any additional terms or conditions.
-      Notwithstanding the above, nothing herein shall supersede or modify
-      the terms of any separate license agreement you may have executed
-      with Licensor regarding such Contributions.
-
-   6. Trademarks. This License does not grant permission to use the trade
-      names, trademarks, service marks, or product names of the Licensor,
-      except as required for reasonable and customary use in describing the
-      origin of the Work and reproducing the content of the NOTICE file.
-
-   7. Disclaimer of Warranty. Unless required by applicable law or
-      agreed to in writing, Licensor provides the Work (and each
-      Contributor provides its Contributions) on an "AS IS" BASIS,
-      WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
-      implied, including, without limitation, any warranties or conditions
-      of TITLE, NON-INFRINGEMENT, MERCHANTABILITY, or FITNESS FOR A
-      PARTICULAR PURPOSE. You are solely responsible for determining the
-      appropriateness of using or redistributing the Work and assume any
-      risks associated with Your exercise of permissions under this License.
-
-   8. Limitation of Liability. In no event and under no legal theory,
-      whether in tort (including negligence), contract, or otherwise,
-      unless required by applicable law (such as deliberate and grossly
-      negligent acts) or agreed to in writing, shall any Contributor be
-      liable to You for damages, including any direct, indirect, special,
-      incidental, or consequential damages of any character arising as a
-      result of this License or out of the use or inability to use the
-      Work (including but not limited to damages for loss of goodwill,
-      work stoppage, computer failure or malfunction, or any and all
-      other commercial damages or losses), even if such Contributor
-      has been advised of the possibility of such damages.
-
-   9. Accepting Warranty or Additional Liability. While redistributing
-      the Work or Derivative Works thereof, You may choose to offer,
-      and charge a fee for, acceptance of support, warranty, indemnity,
-      or other liability obligations and/or rights consistent with this
-      License. However, in accepting such obligations, You may act only
-      on Your own behalf and on Your sole responsibility, not on behalf
-      of any other Contributor, and only if You agree to indemnify,
-      defend, and hold each Contributor harmless for any liability
-      incurred by, or claims asserted against, such Contributor by reason
-      of your accepting any such warranty or additional liability.
-
-   END OF TERMS AND CONDITIONS
-
-   APPENDIX: How to apply the Apache License to your work.
-
-      To apply the Apache License to your work, attach the following
-      boilerplate notice, with the fields enclosed by brackets "[]"
-      replaced with your own identifying information. (Don't include
-      the brackets!)  The text should be enclosed in the appropriate
-      comment syntax for the file format. We also recommend that a
-      file or class name and description of purpose be included on the
-      same "printed page" as the copyright notice for easier
-      identification within third-party archives.
-
-   Copyright [yyyy] [name of copyright owner]
-
-   Licensed under the Apache License, Version 2.0 (the "License");
-   you may not use this file except in compliance with the License.
-   You may obtain a copy of the License at
-
-       http://www.apache.org/licenses/LICENSE-2.0
-
-   Unless required by applicable law or agreed to in writing, software
-   distributed under the License is distributed on an "AS IS" BASIS,
-   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-   See the License for the specific language governing permissions and
-   limitations under the License.
-
-*/
-
-import { expect } from 'chai'
-import { ethers } from 'hardhat'
-import { providers } from 'ethers'
-import { SignerWithAddress } from '@nomiclabs/hardhat-ethers/signers.js'
-import { time } from '@nomicfoundation/hardhat-network-helpers'
-import { isinGenerator } from '@thomaschaplin/isin-generator'
-import {
-    type ResolverProxy,
-    type AdjustBalances,
-    type ERC1410ScheduledTasks,
-    type AccessControl,
-    TimeTravel,
-    Lock,
-    Equity,
-    Cap,
-    IFactory,
-    BusinessLogicResolver,
-    AccessControl__factory,
-    ERC1410ScheduledTasks__factory,
-    AdjustBalances__factory,
-    Cap__factory,
-    Lock__factory,
-    Equity__factory,
-    TimeTravel__factory,
-    Kyc,
-    SsiManagement,
-} from '@typechain'
-import {
-    ADJUSTMENT_BALANCE_ROLE,
-    PAUSER_ROLE,
-    ISSUER_ROLE,
-    CAP_ROLE,
-    CONTROLLER_ROLE,
-    LOCKER_ROLE,
-    KYC_ROLE,
-    SSI_MANAGER_ROLE,
-    CORPORATE_ACTION_ROLE,
-    deployEquityFromFactory,
-    Rbac,
-    RegulationSubType,
-    RegulationType,
-    DeployAtsFullInfrastructureCommand,
-    deployAtsFullInfrastructure,
-    MAX_UINT256,
-    ZERO,
-    EMPTY_STRING,
-} from '@scripts'
-import { dateToUnixTimestamp } from '../../../dateFormatter'
-
-const amount = 1
-const balanceOf_A_Original = [10 * amount, 100 * amount]
-const balanceOf_B_Original = [20 * amount, 200 * amount]
-const _PARTITION_ID_1 =
-    '0x0000000000000000000000000000000000000000000000000000000000000001'
-const _PARTITION_ID_2 =
-    '0x0000000000000000000000000000000000000000000000000000000000000002'
-const adjustFactor = 253
-const adjustDecimals = 2
-const decimals_Original = 6
-const maxSupply_Original = 1000000 * amount
-const maxSupply_Partition_1_Original = 50000 * amount
-const maxSupply_Partition_2_Original = 0
-const ONE_SECOND = 1
-const EMPTY_VC_ID = EMPTY_STRING
-
-describe('Locks Layer 2 Tests', () => {
-    let diamond: ResolverProxy
-    let signer_A: SignerWithAddress
-    let signer_B: SignerWithAddress
-    let signer_C: SignerWithAddress
-
-    let account_A: string
-    let account_B: string
-    let account_C: string
-
-    let factory: IFactory
-    let businessLogicResolver: BusinessLogicResolver
-    let erc1410Facet: ERC1410ScheduledTasks
-    let adjustBalancesFacet: AdjustBalances
-    let accessControlFacet: AccessControl
-    let capFacet: Cap
-    let equityFacet: Equity
-    let lockFacet: Lock
-    let timeTravelFacet: TimeTravel
-    let kycFacet: Kyc
-    let ssiManagementFacet: SsiManagement
-
-    async function deployAsset(multiPartition: boolean) {
-        const init_rbacs: Rbac[] = set_initRbacs()
-
-        diamond = await deployEquityFromFactory({
-            adminAccount: account_A,
-            isWhiteList: false,
-            isControllable: true,
-            arePartitionsProtected: false,
-            isMultiPartition: multiPartition,
-            name: 'TEST_AccessControl',
-            symbol: 'TAC',
-            decimals: decimals_Original,
-            isin: isinGenerator(),
-            votingRight: false,
-            informationRight: false,
-            liquidationRight: false,
-            subscriptionRight: true,
-            conversionRight: true,
-            redemptionRight: true,
-            putRight: false,
-            dividendRight: 1,
-            currency: '0x345678',
-            numberOfShares: MAX_UINT256,
-            nominalValue: 100,
-            regulationType: RegulationType.REG_D,
-            regulationSubType: RegulationSubType.REG_D_506_B,
-            countriesControlListType: true,
-            listOfCountries: 'ES,FR,CH',
-            info: 'nothing',
-            init_rbacs,
-            factory,
-            businessLogicResolver: businessLogicResolver.address,
-        })
-
-        await setFacets({ diamond, defaultSigner: signer_A })
-    }
-
-    async function setFacets({
-        diamond,
-        defaultSigner = ethers.provider,
-    }: {
-        diamond: ResolverProxy
-        defaultSigner?: SignerWithAddress | providers.Provider
-    }) {
-        accessControlFacet = AccessControl__factory.connect(
-            diamond.address,
-            defaultSigner
-        )
-        erc1410Facet = ERC1410ScheduledTasks__factory.connect(
-            diamond.address,
-            defaultSigner
-        )
-        adjustBalancesFacet = AdjustBalances__factory.connect(
-            diamond.address,
-            defaultSigner
-        )
-        capFacet = Cap__factory.connect(diamond.address, defaultSigner)
-        lockFacet = Lock__factory.connect(diamond.address, defaultSigner)
-        equityFacet = Equity__factory.connect(diamond.address, defaultSigner)
-        timeTravelFacet = TimeTravel__factory.connect(
-            diamond.address,
-            defaultSigner
-        )
-        kycFacet = await ethers.getContractAt('Kyc', diamond.address, signer_B)
-        ssiManagementFacet = await ethers.getContractAt(
-            'SsiManagement',
-            diamond.address,
-            signer_A
-        )
-    }
-
-    function set_initRbacs(): Rbac[] {
-        const rbacPause: Rbac = {
-            role: PAUSER_ROLE,
-            members: [account_B],
-        }
-        const corporateActionPause: Rbac = {
-            role: CORPORATE_ACTION_ROLE,
-            members: [account_B],
-        }
-        const rbacKYC: Rbac = {
-            role: KYC_ROLE,
-            members: [account_B],
-        }
-        const rbacSSI: Rbac = {
-            role: SSI_MANAGER_ROLE,
-            members: [account_A],
-        }
-        return [rbacPause, corporateActionPause, rbacKYC, rbacSSI]
-    }
-
-    async function setPreBalanceAdjustment() {
-        // Granting Role to account C
-        accessControlFacet = accessControlFacet.connect(signer_A)
-        await accessControlFacet.grantRole(ADJUSTMENT_BALANCE_ROLE, account_C)
-        await accessControlFacet.grantRole(ISSUER_ROLE, account_A)
-        await accessControlFacet.grantRole(CAP_ROLE, account_A)
-        await accessControlFacet.grantRole(CONTROLLER_ROLE, account_A)
-        await accessControlFacet.grantRole(LOCKER_ROLE, account_A)
-
-        // Using account C (with role)
-        adjustBalancesFacet = adjustBalancesFacet.connect(signer_C)
-        erc1410Facet = erc1410Facet.connect(signer_A)
-        capFacet = capFacet.connect(signer_A)
-
-        await capFacet.setMaxSupply(maxSupply_Original)
-        await capFacet.setMaxSupplyByPartition(
-            _PARTITION_ID_1,
-            maxSupply_Partition_1_Original
-        )
-        await capFacet.setMaxSupplyByPartition(
-            _PARTITION_ID_2,
-            maxSupply_Partition_2_Original
-        )
-
-        await ssiManagementFacet.connect(signer_A).addIssuer(account_A)
-        await kycFacet.grantKyc(
-            account_A,
-            EMPTY_VC_ID,
-            ZERO,
-            MAX_UINT256,
-            account_A
-        )
-        await kycFacet.grantKyc(
-            account_B,
-            EMPTY_VC_ID,
-            ZERO,
-            MAX_UINT256,
-            account_A
-        )
-        await erc1410Facet.issueByPartition({
-            partition: _PARTITION_ID_1,
-            tokenHolder: account_A,
-            value: balanceOf_A_Original[0],
-            data: '0x',
-        })
-        await erc1410Facet.issueByPartition({
-            partition: _PARTITION_ID_2,
-            tokenHolder: account_A,
-            value: balanceOf_A_Original[1],
-            data: '0x',
-        })
-        await erc1410Facet.issueByPartition({
-            partition: _PARTITION_ID_1,
-            tokenHolder: account_B,
-            value: balanceOf_B_Original[0],
-            data: '0x',
-        })
-        await erc1410Facet.issueByPartition({
-            partition: _PARTITION_ID_2,
-            tokenHolder: account_B,
-            value: balanceOf_B_Original[1],
-            data: '0x',
-        })
-    }
-
-    before(async () => {
-        // mute | mock console.log
-        console.log = () => {}
-        // eslint-disable-next-line @typescript-eslint/no-extra-semi
-        ;[signer_A, signer_B, signer_C] = await ethers.getSigners()
-        account_A = signer_A.address
-        account_B = signer_B.address
-        account_C = signer_C.address
-
-        const { deployer, ...deployedContracts } =
-            await deployAtsFullInfrastructure(
-                await DeployAtsFullInfrastructureCommand.newInstance({
-                    signer: signer_A,
-                    useDeployed: false,
-                    useEnvironment: true,
-                    timeTravelEnabled: true,
-                })
-            )
-
-        factory = deployedContracts.factory.contract
-        businessLogicResolver = deployedContracts.businessLogicResolver.contract
-    })
-
-    beforeEach(async () => {
-        await deployAsset(true)
-    })
-
-    afterEach(async () => {
-        await timeTravelFacet.resetSystemTimestamp()
-    })
-
-    it('GIVEN a lock WHEN adjustBalances THEN lock amount gets updated succeeds', async () => {
-        await setPreBalanceAdjustment()
-
-        const balance_Before = await erc1410Facet.balanceOf(account_A)
-        const balance_Before_Partition_1 =
-            await erc1410Facet.balanceOfByPartition(_PARTITION_ID_1, account_A)
-
-        // LOCK
-        lockFacet = lockFacet.connect(signer_A)
-        await lockFacet.lockByPartition(
-            _PARTITION_ID_1,
-            amount,
-            account_A,
-            dateToUnixTimestamp('2030-01-01T00:00:01Z')
-        )
-
-        const lock_TotalAmount_Before = await lockFacet.getLockedAmountFor(
-            account_A
-        )
-        const lock_TotalAmount_Before_Partition_1 =
-            await lockFacet.getLockedAmountForByPartition(
-                _PARTITION_ID_1,
-                account_A
-            )
-        const lock_Before = await lockFacet.getLockForByPartition(
-            _PARTITION_ID_1,
-            account_A,
-            1
-        )
-
-        // adjustBalances
-        await adjustBalancesFacet.adjustBalances(adjustFactor, adjustDecimals)
-
-        // scheduled two balance updates
-        equityFacet = equityFacet.connect(signer_B)
-
-        const balanceAdjustmentData = {
-            executionDate: dateToUnixTimestamp(
-                '2030-01-01T00:00:02Z'
-            ).toString(),
-            factor: adjustFactor,
-            decimals: adjustDecimals,
-        }
-
-        const balanceAdjustmentData_2 = {
-            executionDate: dateToUnixTimestamp(
-                '2030-01-01T00:16:40Z'
-            ).toString(),
-            factor: adjustFactor,
-            decimals: adjustDecimals,
-        }
-        await equityFacet.setScheduledBalanceAdjustment(balanceAdjustmentData)
-        await equityFacet.setScheduledBalanceAdjustment(balanceAdjustmentData_2)
-
-        // wait for first scheduled balance adjustment only
-        await timeTravelFacet.changeSystemTimestamp(
-            dateToUnixTimestamp('2030-01-01T00:00:03Z')
-        )
-
-        const lock_TotalAmount_After = await lockFacet.getLockedAmountFor(
-            account_A
-        )
-        const lock_TotalAmount_After_Partition_1 =
-            await lockFacet.getLockedAmountForByPartition(
-                _PARTITION_ID_1,
-                account_A
-            )
-        const lock_After = await lockFacet.getLockForByPartition(
-            _PARTITION_ID_1,
-            account_A,
-            1
-        )
-        const balance_After = await erc1410Facet.balanceOf(account_A)
-        const balance_After_Partition_1 =
-            await erc1410Facet.balanceOfByPartition(_PARTITION_ID_1, account_A)
-
-        expect(lock_TotalAmount_After).to.be.equal(
-            lock_TotalAmount_Before.mul(adjustFactor * adjustFactor)
-        )
-        expect(lock_TotalAmount_After_Partition_1).to.be.equal(
-            lock_TotalAmount_Before_Partition_1.mul(adjustFactor * adjustFactor)
-        )
-        expect(balance_After).to.be.equal(
-            balance_Before.sub(amount).mul(adjustFactor * adjustFactor)
-        )
-        expect(lock_TotalAmount_After).to.be.equal(
-            lock_TotalAmount_Before.mul(adjustFactor * adjustFactor)
-        )
-        expect(balance_After_Partition_1).to.be.equal(
-            balance_Before_Partition_1
-                .sub(amount)
-                .mul(adjustFactor * adjustFactor)
-        )
-        expect(lock_After.amount_).to.be.equal(
-            lock_Before.amount_.mul(adjustFactor * adjustFactor)
-        )
-    })
-
-    it('GIVEN a lock WHEN adjustBalances THEN release succeeds', async () => {
-        await setPreBalanceAdjustment()
-        const balance_Before = await erc1410Facet.balanceOf(account_A)
-        const balance_Before_Partition_1 =
-            await erc1410Facet.balanceOfByPartition(_PARTITION_ID_1, account_A)
-
-        // LOCK TWICE
-        const currentTimestamp = (await ethers.provider.getBlock('latest'))
-            .timestamp
-
-        lockFacet = lockFacet.connect(signer_A)
-        await lockFacet.lockByPartition(
-            _PARTITION_ID_1,
-            amount,
-            account_A,
-            currentTimestamp + ONE_SECOND
-        )
-        await lockFacet.lockByPartition(
-            _PARTITION_ID_1,
-            amount,
-            account_A,
-            currentTimestamp + 100 * ONE_SECOND
-        )
-
-        const locked_Amount_Before = await lockFacet.getLockedAmountFor(
-            account_A
-        )
-        const locked_Amount_Before_Partition_1 =
-            await lockFacet.getLockedAmountForByPartition(
-                _PARTITION_ID_1,
-                account_A
-            )
-
-        // adjustBalances
-        await adjustBalancesFacet.adjustBalances(adjustFactor, adjustDecimals)
-
-        // RELEASE LOCK
-        await time.setNextBlockTimestamp(
-            (await ethers.provider.getBlock('latest')).timestamp +
-                2 * ONE_SECOND
-        )
-        await lockFacet.releaseByPartition(_PARTITION_ID_1, 1, account_A)
-
-        const balance_After_Release = await erc1410Facet.balanceOf(account_A)
-        const balance_After_Release_Partition_1 =
-            await erc1410Facet.balanceOfByPartition(_PARTITION_ID_1, account_A)
-        const locked_Amount_After = await lockFacet.getLockedAmountFor(
-            account_A
-        )
-        const locked_Amount_After_Partition_1 =
-            await lockFacet.getLockedAmountForByPartition(
-                _PARTITION_ID_1,
-                account_A
-            )
-
-        expect(balance_After_Release).to.be.equal(
-            balance_Before.sub(amount).mul(adjustFactor)
-        )
-        expect(balance_After_Release_Partition_1).to.be.equal(
-            balance_Before_Partition_1.sub(amount).mul(adjustFactor)
-        )
-        expect(locked_Amount_After).to.be.equal(
-            locked_Amount_Before.sub(amount).mul(adjustFactor)
-        )
-        expect(locked_Amount_After_Partition_1).to.be.equal(
-            locked_Amount_Before_Partition_1.sub(amount).mul(adjustFactor)
-        )
-        expect(balance_After_Release.add(locked_Amount_After)).to.be.equal(
-            balance_Before.mul(adjustFactor)
-        )
-        expect(
-            balance_After_Release_Partition_1.add(
-                locked_Amount_After_Partition_1
-            )
-        ).to.be.equal(balance_Before_Partition_1.mul(adjustFactor))
-    })
-
-    it('GIVEN a lock WHEN adjustBalances THEN lock succeeds', async () => {
-        await setPreBalanceAdjustment()
-        const balance_Before = await erc1410Facet.balanceOf(account_A)
-        const balance_Before_Partition_1 =
-            await erc1410Facet.balanceOfByPartition(_PARTITION_ID_1, account_A)
-
-        // LOCK BEFORE BALANCE ADJUSTMENT
-        const currentTimestamp = (await ethers.provider.getBlock('latest'))
-            .timestamp
-
-        lockFacet = lockFacet.connect(signer_A)
-        await lockFacet.lockByPartition(
-            _PARTITION_ID_1,
-            amount,
-            account_A,
-            currentTimestamp + 100 * ONE_SECOND
-        )
-
-        const locked_Amount_Before = await lockFacet.getLockedAmountFor(
-            account_A
-        )
-        const locked_Amount_Before_Partition_1 =
-            await lockFacet.getLockedAmountForByPartition(
-                _PARTITION_ID_1,
-                account_A
-            )
-
-        // adjustBalances
-        await adjustBalancesFacet.adjustBalances(adjustFactor, adjustDecimals)
-
-        // LOCK AFTER BALANCE ADJUSTMENT
-        lockFacet = lockFacet.connect(signer_A)
-        await lockFacet.lockByPartition(
-            _PARTITION_ID_1,
-            amount,
-            account_A,
-            currentTimestamp + 100 * ONE_SECOND
-        )
-
-        const balance_After_Lock = await erc1410Facet.balanceOf(account_A)
-        const balance_After_Lock_Partition_1 =
-            await erc1410Facet.balanceOfByPartition(_PARTITION_ID_1, account_A)
-        const locked_Amount_After = await lockFacet.getLockedAmountFor(
-            account_A
-        )
-        const locked_Amount_After_Partition_1 =
-            await lockFacet.getLockedAmountForByPartition(
-                _PARTITION_ID_1,
-                account_A
-            )
-
-        expect(balance_After_Lock).to.be.equal(
-            balance_Before.sub(amount).mul(adjustFactor).sub(amount)
-        )
-        expect(balance_After_Lock_Partition_1).to.be.equal(
-            balance_Before_Partition_1.sub(amount).mul(adjustFactor).sub(amount)
-        )
-        expect(locked_Amount_After).to.be.equal(
-            locked_Amount_Before.mul(adjustFactor).add(amount)
-        )
-        expect(locked_Amount_After_Partition_1).to.be.equal(
-            locked_Amount_Before_Partition_1.mul(adjustFactor).add(amount)
-        )
-        expect(balance_After_Lock.add(locked_Amount_After)).to.be.equal(
-            balance_Before.mul(adjustFactor)
-        )
-        expect(
-            balance_After_Lock_Partition_1.add(locked_Amount_After_Partition_1)
-        ).to.be.equal(balance_Before_Partition_1.mul(adjustFactor))
-    })
-
-    it('GIVEN a lock WHEN adjustBalances and release THEN reset lock LABAF succeeds', async () => {
-        await setPreBalanceAdjustment()
-
-        // LOCK
-        const currentTimestamp = (await ethers.provider.getBlock('latest'))
-            .timestamp
-
-        lockFacet = lockFacet.connect(signer_A)
-        await lockFacet.lockByPartition(
-            _PARTITION_ID_1,
-            amount,
-            account_A,
-            currentTimestamp + ONE_SECOND
-        )
-        const TotalLockLABAF_Before =
-            await adjustBalancesFacet.getTotalLockLabafByPartition(
-                _PARTITION_ID_1,
-                account_A
-            )
-
-        // adjustBalances
-        await adjustBalancesFacet.adjustBalances(adjustFactor, adjustDecimals)
-
-        // LOCK AGAIN
-        await lockFacet.lockByPartition(
-            _PARTITION_ID_1,
-            amount,
-            account_A,
-            currentTimestamp + 100 * ONE_SECOND
-        )
-
-        const LockLABAF_Before =
-            await adjustBalancesFacet.getLockLabafByPartition(
-                _PARTITION_ID_1,
-                2,
-                account_A
-            )
-
-        // RELEASE LOCK
-        await time.setNextBlockTimestamp(
-            (await ethers.provider.getBlock('latest')).timestamp +
-                2 * ONE_SECOND
-        )
-        await lockFacet.releaseByPartition(_PARTITION_ID_1, 1, account_A)
-
-        const TotalLockLABAF_After =
-            await adjustBalancesFacet.getTotalLockLabafByPartition(
-                _PARTITION_ID_1,
-                account_A
-            )
-        const LockLABAF_After =
-            await adjustBalancesFacet.getLockLabafByPartition(
-                _PARTITION_ID_1,
-                2,
-                account_A
-            )
-
-        expect(LockLABAF_Before).to.be.equal(LockLABAF_After)
-        expect(TotalLockLABAF_Before).to.be.equal(0)
-        expect(TotalLockLABAF_After).to.be.equal(adjustFactor)
-        expect(LockLABAF_Before).to.be.equal(adjustFactor)
-    })
-})
